aiSearchPython/requirements.txt
semantic-kernel
python-dotenv
azure-core
azure-search
# azure-search-documents
azure-identity
tenacity
azure-search-documents==11.4.0b8
numpy
pandas
plotly
streamlit
openai
SQLAlchemy
pyarrow
scikit-learn
pmdarima
matplotlib
pyodbc